Split `base64.h` into `.h` / `.c` files
